+++ packages/templates/packages.html
@@ -11,7 +11,7 @@
       <div class="modal-dialog  modal-dialog-centered">
         <div class="modal-content">
           <div class="modal-header">
-            <h5 class="modal-title" id="trackidDelModalLabel">Удалить трек-номер <strong><span id="deltrackid"></span></strong></h5>
+            <h5 class="modal-title" id="trackidDelModalLabel">Удалить трек-номер <strong><span id="deltrackid"></span></strong>?</h5>
             <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
           </div>
           <div class="modal-body">
@@ -34,15 +34,15 @@
         </div>
         {% endif %}
         {% for package in packages %}
-        <div class="card mt-2" id="track_{{ package.package_id }}">
+        <div class="card mt-2" id="track_{{ package.package_id.id }}">
             <div class="card-body">
-                <button type="button" class="btn-close float-end  btn-del" onclick="DeleteTrackId('{{ package.package_id }}')" aria-label="Удалить"></button>
+                <button type="button" class="btn-close float-end  btn-del" onclick="DeleteTrackId('{{ package.package_id.id }}')" aria-label="Удалить"></button>
             {% if package.desc %}
                 <h3 class="card-title">{{ package.desc }}</h3>
                 <h6 class="card-subtitle mb-2 text-body-secondary">{{ package.package_id.get_status_display }} {{ package.package_id.status_change_date|date:"d.m.Y H:i" }}</h6>
-                <p class="card-text">Трек-номер: {{ package.package_id }}</p>
+                <p class="card-text">Трек-номер: {{ package.package_id.id }}</p>
             {% else %}
-                <h3 class="card-title">{{ package.package_id }}</h3>
+                <h3 class="card-title">{{ package.package_id.id }}</h3>
                 <h6 class="card-subtitle mb-2 text-body-secondary">{{ package.package_id.get_status_display }} {{ package.package_id.status_change_date|date:"d.m.Y H:i" }}</h6>
             {% endif %}
             </div>
